requirements.txt
streamlit
replicate
PyPDF2
pinecone-client
pypdf
langchain
pdf2image
pytesseract
Pillow 
sentence-transformers
pdfplumber
